+++ lib/features/filter_editor/filter_editor.dart
@@ -13,6 +13,7 @@ import '/core/platform/io/io_helper.dart';
 import '/features/filter_editor/widgets/filter_editor_appbar.dart';
 import '/pro_image_editor.dart';
 import '/shared/services/content_recorder/widgets/content_recorder.dart';
+import '/shared/utils/file_constructor_utils.dart';
 import '/shared/widgets/layer/layer_stack.dart';
 import '/shared/widgets/transform/transformed_content_generator.dart';
 import 'types/filter_matrix.dart';
@@ -67,7 +68,7 @@ class FilterEditor extends StatefulWidget
   }) {
     return FilterEditor._(
       key: key,
-      editorImage: EditorImage(file: file),
+      editorImage: EditorImage(file: ensureFileInstance(file)),
       initConfigs: initConfigs,
     );
   }
@@ -120,7 +121,7 @@ class FilterEditor extends StatefulWidget
       );
     } else if (file != null || editorImage?.file != null) {
       return FilterEditor.file(
-        file ?? editorImage!.file!,
+        ensureFileInstance(file ?? editorImage!.file!),
         key: key,
         initConfigs: initConfigs,
       );